refactor: audit code
